% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binbeta.R
\name{binbeta}
\alias{binbeta}
\title{Binomial sampling distribution with a beta distribution of prior}
\usage{
binbeta(alpha = 1, beta = 1, n, r, theta = seq(0, 1, 0.001))
}
\arguments{
\item{alpha}{parameter for the beta distribution ( \eqn{\ge 0} ).}

\item{beta}{parameter for the beta distribution ( \eqn{\ge 0} ).}

\item{n}{the number of trials in binomial distribution.}

\item{r}{the number of successes in n trials.}

\item{\eqn{\theta}}{the range of the probability of success.}
}
\value{
An object of class "\code{binbeta}" is returned.
\item{prior}{the prior distribution, i.e.
the \eqn{beta(\alpha,\beta)} distribution} \item{likelihood}{the likelihood
function of r given \eqn{\theta}, i.e. the
\eqn{binomial(n,r)} distribution} \item{posterior}{the
posterior distribution of \eqn{\theta} given r, i.e. the
\eqn{beta(\alpha+r, \beta+n-r)} distribution} \item{theta}{the
range of the probability of success} \item{pri.alpha}{the
alpha parameter for the beta distribution of prior} \item{pri.beta}{the
beta parameter for the beta distribution of prior} \item{pos.alpha}{the
alpha parameter for the beta distribution of posterior} \item{pos.beta}{the
beta parameter for the beta distribution of posterior}
}
\description{
Define the posterior distribution function for \eqn{\pi ( \theta | r )}, with
a beta prior distribution \eqn{\pi ( \theta; a, b )} and a binomial sampling
distribution \eqn{p ( r | \theta )}.
}
\examples{
## simplest one with 3 successes in 10 trials and a uniform beta(1,1) prior
binbeta(n = 6, r = 2)

## example 3.6 : 3 successes in 10 trials and a beta(4,6) prior
binbeta(4, 6, 10, 3)

}
\seealso{
\code{\link{summary.binbeta}} will return summararies of prior
and posterior distribution.
}
