% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamgam.R
\name{gamgam}
\alias{gamgam}
\title{Gamma sampling distribution with a gamma distribution of prior}
\source{
For theory details, based on
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{Wikipedia's article on
gamma conjugate prior} and STATG012 slides5 Example5.4 on Moodle at UCL.
\href{https://moodle.ucl.ac.uk/mod/folder/view.php?id=2570901}{STATG012
slides5}
}
\usage{
gamgam(t, a = 1, shape, rate, scale = 1/rate)
}
\arguments{
\item{t}{a sample data from a gamma distribution with known shape a and rate
theta ( t > 0 )}

\item{a}{the known shape parameter of the gamma sampling distribution ( > 0)}

\item{shape}{the shape parameter for the gamma distribution of prior ( > 0).}

\item{rate}{the rate parameter for the gamma distribution of prior( > 0).}

\item{scale}{equals 1 / rate ( > 0).}
}
\value{
An object of class "\code{g12post}" is returned.
\item{prior}{the prior distribution, i.e.
the \eqn{gamma(\alpha,\lambda} distribution.} \item{likelihood}{the likelihood
function of t given \eqn{\theta}, i.e. the
\eqn{gamma(a,\theta)} distribution.} \item{posterior}{the
posterior distribution of \eqn{\theta} given t.} \item{theta}{the
rate parameter of the gamma distribution and its prior is
a gamma distribution
\eqn{\pi ( \theta; \alpha, \lambda )}.} \item{pri.shape}{the
shape parameter for the gamma distribution of prior.} \item{pri.rate}{the
rate parameter for the gamma distribution of prior.} \item{pos.shape}{the
shape parameter for the gamma distribution of
posterior.} \item{pos.rate}{the rate parameter for the gamma distribution
of posterior.}\item{model}{the prior and likelihood type to produce the
posterior.}
}
\description{
Define the posterior distribution function for \eqn{\pi (\theta | t )},
with a gamma prior distribution \eqn{\pi ( \theta; \alpha, \lambda )} and a
gamma sampling distribution with known shape parameter a
\eqn{p (t; a, \theta)}.
}
\examples{
## an exponential distribution with a gamma prior, similiar
## as Example 5.4 from slides 6
## generate a sample of 10 observations from an exponential distribution
x <- rexp(10)
## find the posterior density and summary it
gam <- gamgam(x, a = 1, 4, 2)
summary(gam)

## generate a sample of 50 observations from a gamma distribution with a = 2
y <- rgamma(10, shape = 2)
## find the posterior density and plot it
ex <- gamgam(y, a = 2, 4, 2)
plot(ex, leg_pos = "top", box.col = "n")
}
\seealso{
\code{\link{summary.g12post}} for summararies of prior
and posterior distribution.

\code{\link{plot.g12post}} for plots of prior and posterior
distribution.
}
