% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\title{Gamma sampling distribution with a gamma distribution of prior}
\usage{
gamma(t, a = 1, shape, rate, scale = 1/rate)
}
\arguments{
\item{t}{a sample data from a  gamma distribution with shape a and rate
theta ( t > 0 )}

\item{a}{the shape of the gamma distribution ( > 0)}

\item{shape}{the shape parameter for the gamma distribution of prior ( > 0).}

\item{rate}{the rate parameter for the gamma distribution of prior( > 0).}

\item{scale}{equals 1 / rate ( > 0).}
}
\value{
An object of class "\code{g12post}" is returned.
}
\description{
Define the posterior distribution function for \eqn{\pi (\theta | t )},
with a gamma prior distribution \eqn{\pi ( \thta; \alpha, \lambda )} and a
gamma sampling distribution \eqn{p (t; a, \theta)}.
}
\examples{

## an exponential distribution with a gamma prior, similiar
## as Example 5.4 from slides 6
## generate a sample of 10 observations from an exponential distribution
x <- rexp(10)

## find the posterior density
gamma(x, a = 1, 4, 2)

## generate a sample of 50 observations from a gamma distribution with a = 2
y <- rgamma(50, shape = 2)

## find the posterior density
gamma(y, a = 2, 4, 2)

}
\seealso{
The
\href{https://moodle.ucl.ac.uk/mod/folder/view.php?id=2570901}{slides}
of STATG012 on Moodle
}
