% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalproblem.r
\name{normalprob}
\alias{normalprob}
\title{show problem of normal when mu and sigma are unknow}
\usage{
normalprob(x, m, s, alpha, beta)
}
\arguments{
\item{m}{and s are parameters of mu which follows normal distribution}

\item{alpha}{and beta are parameters of tau(1/sigma^2) which follows gamma
distribution}
}
\description{
show problem of normal when mu and sigma are unknow
}
\examples{
##
x <- rnorm(10, sd = 1)
a <- normalprob(x, m = 2, s = 1, a = 2, b = 2)

## problem1: the posterior is almost zero
a$posterior

## problem2 : the mu.prior distribution is more concentrated than the mu.posterior,
##for tau is fine, but for all prior and posterior is not, because posterior is 0.
##black line is prior and red line is posterior
muplot <- matplot(cbind(a$mu,a$mu),cbind(a$mu.prior,a$mu.pos),col = 1:2)
tauplot <- matplot(,cbind(a$tau.prior,a$tau.pos),col = 1:2)
all <- matplot(cbind(a$prior,a$posterior),col = 1:2)

}
