% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_test.R
\name{normaltest}
\alias{normaltest}
\title{Bayesian hypothesis testing with Normal distribution}
\source{
The
\href{https://moodle.ucl.ac.uk/mod/folder/view.php?id=2570901}{slides 9}
of STATG012 on Moodle
}
\usage{
normaltest(x, theta, a, b)
}
\arguments{
\item{x}{results from the \code{\link{normnorm}} with known variance.}

\item{theta}{the parameter for hypothesis testing.}

\item{a}{loss due to type I error.}

\item{b}{loss due to typp II error.}
}
\value{
hypothesis testing result and expected posterior loss
}
\description{
The Bayesian hypothesis testing is based on the posterior
distribution \eqn{\pi(\theta | x)}, and the decision is
to reject or accept the null hypothesis according to
which decision provides the smaller losses. The loss of
rejecting the null hypothesis is a times the probability
of the null is true, where a is the loss due to type I error.
The loss of accepting the null hypothesis is b times the probability
of the null is false, where b is the loss due to type II error.
}
\examples{
## generate a sample of 16 from a normal distribution with sd=4
y <- rnorm(16, mean = 5.2, sd=4)
## the observed sample mean is 5.2
yy <- y- mean(y)+5.2
## find the posterior density
exmp1 <- normnorm(yy, m = 4.5, s = sqrt(10), sigma = 4)
## make the hypothesis testing
normaltest(exmp1, 5, 1, 1)
}
