% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normnorm.R
\name{normnorm}
\alias{normnorm}
\title{Normal sampling distribution with a normal distribution of prior}
\usage{
normnorm(x, m, s, alpha = NULL, beta = NULL, mu = NULL, sigma = NULL)
}
\arguments{
\item{x}{a vector of observations from a normal distribution with mu and
sigma.}

\item{mu}{the mean of x vector from a normal distribution. If it is NULL,
the mean of x is unknown.}

\item{sigma}{the standard deviation of x vector from a normal distribution.
If it is NULL, the standard deviation of x is unknown.}

\item{\eqn{m}, \eqn{s}}{parameters from a normal distribution of mu prior.}

\item{\eqn{\alpha}, \eqn{\beta}}{parameters from a gamma distribution. If
mu and sigma is NULL, the prior distribution on \eqn{\mu} and precision
\eqn{\tau} has a Normal-Gamma distribution. \eqn{\tau = 1/\sigma^2}}
}
\value{
An object of class "\code{g12post}" is returned.
}
\description{
Define the posterior distribution function for \eqn{f( x | \mu, \sigma )}, with
a normal prior distribution \eqn{g( \mu, \sigma )} and a normal sampling
distribution \eqn{f( \mu, \sigma | x )}.
}
\examples{

## this example is from slides6 Example 5.7
## generate a sample of 9 from a normal distribution with sd=2
x <- rnorm(9, sd = 2)
## the observed sample mean is 20
xx <- x- mean(x) + 20
## find the posterior density
normnorm(xx, m = 25, s = sqrt(10), sigma = 2)

}
\seealso{
The
\href{https://moodle.ucl.ac.uk/mod/folder/view.php?id=2570901}{slides}
of STATG012 on Moodle
}
