% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisgamma.R
\name{poisgamma}
\alias{poisgamma}
\title{Poisson sampling distribution with a gamma distribution of prior}
\source{
STATG012 slides5 Example5.6 on Moodle at UCL
\href{https://moodle.ucl.ac.uk/mod/folder/view.php?id=2570901}{STATG012
slides5}
}
\usage{
poisgamma(x, shape, rate, scale = 1/rate)
}
\arguments{
\item{x}{the sample data from poisson distribution  (x > 0).}

\item{shape}{the shape parameter for a gamma distribution (> 0).}

\item{rate}{the rate parameter for a gamma distribution (> 0).}

\item{scale}{equals 1 / rate (> 0).}
}
\value{
An object of class "\code{g12post}" is returned.
\item{prior}{the prior distribution, i.e.
the \eqn{gamma(\alpha,\lambda)} distribution.} \item{likelihood}{the
likelihood function of x given \eqn{\mu}, i.e. the
\eqn{Poisson( x | \mu)} distribution.} \item{posterior}{the
posterior distribution of \eqn{\mu} given x.} \item{mu}{the expected number
of occurence which is the parameter of poisson
distribution.} \item{pri.shape}{the shape parameter for the gamma
distribution of prior.} \item{pri.rate}{the rate parameter for
the gamma distribution of prior.} \item{pos.shape}{the shape parameter for the gamma
distribution of posterior.} \item{pos.rate}{the
rate parameter for the gamma distribution of posterior.}\item{model}{the
prior and likelihood type to produce the posterior.}
}
\description{
Define the posterior distribution function for \eqn{\pi ( \mu | x )}, with
a gamma prior distribution \eqn{\pi ( \mu; \alpha, \lambda )} and a
poisson sampling distribution \eqn{f (x | \mu)}.
}
\examples{
## an observation of 10 with an exponential prior
a <- poisgamma(10, 1, 1)
summary(a)

## An Example 5.6 from slides 6, summary and plot it
x <- c(2, 10)
ex <- poisgamma(x, 4, 1)
summary(ex)
plot(ex,leg_pos = "right",cex = 1, lty = 5:6,
main = "Distributions of Example5.6")

}
\references{
Bolstad, W.M. 2007. Introduction to Bayesian Statistics. (2nd ed.).
Hoboken, New Jersey: John Wiley & Sons, Inc.
}
\seealso{
\code{\link{summary.g12post}} for summararies of prior
and posterior distribution.

\code{\link{plot.g12post}} for plots of prior and posterior
distribution.
}
