% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Creat the posterior distribution}
\usage{
posterior(a, b, c, d, model = "")
}
\arguments{
\item{a}{paramter of the prior distribution}

\item{b}{paramter of the prior distribution}
}
\description{
Creat the posterior distribution
}
