% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statg012.R
\docType{package}
\name{statg012}
\alias{statg012}
\alias{statg012-package}
\title{statg012: A package for Statistical Inference at University College London}
\description{
Package provides R datasets and functions to help students from STATG012
Statistical Inference better understand baysiean inference, compare statistics
between prior density and posterior density and see how R can be utilized to
plot densities of prior and posterior.
}
\section{Please note}{

The main sources of information about the material in STATG012
are the lecture slides with tutorial sheets, text reading etc available
from the
\href{https://moodle.ucl.ac.uk/course/view.php?id=7610}{STAT012 Moodle page}.
The focus of this R package is mainly on R code.
}

