% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_g12post.R
\name{summary.g12post}
\alias{summary.g12post}
\title{Object Summaries}
\usage{
\method{summary}{g12post}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
the function \code{summary.g12post} returns a list of summary
statistics such as mean, variance and quantiles of prior and posterior
distribution, repectively.
}
\description{
A generic function used to produce result summaries of the results of various
model fitthing functions. The distinctive models are \code{\link{binombeta}},
\code{\link{nbinombeta}}, \code{\link{poisgamma}}, \code{\link{gamgam}} and
\code{\link{normnorm}} respectively.
}
\examples{
## Obtain the posterior distribution with binomial-beta model and summary
## it
ep1 <- binombeta(alpha = 1, beta = 2, n = 10, r = 3)
summary(ep1)
}
