% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_test.R
\name{normaltest}
\alias{normaltest}
\title{Bayesian hypothesis testing with Normal distribution}
\usage{
normaltest(x, theta, a, b)
}
\arguments{
\item{theta}{the parameter for hypothesis testing.}

\item{a}{type I error}

\item{b}{typp II error}
}
\description{
Bayesian hypothesis testing with Normal distribution
}
