% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_g12post.R
\name{plot.g12post}
\alias{plot.g12post}
\title{Plot shows the prior and posterior distribution}
\usage{
\method{plot}{g12post}(x, y, leg_pos = c("topright", "bottomright", "bottom",
  "bottomleft", "left", "topleft", "top", "right", "center"), ...)
}
\arguments{
\item{x, y}{the vectors or matrices of data for ploting}

\item{leg_pos}{the place to put the legend}

\item{model}{"binbeta", "poigamma" ...}
}
\description{
Plot shows the prior and posterior distribution
}
