% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot.normal}
\alias{plot.normal}
\title{plot of normal}
\usage{
\method{plot}{normal}(x, y, leg_pos = c("topright", "bottomright", "bottom",
  "bottomleft", "left", "topleft", "top", "right", "center"), which = c(1:4),
  ...)
}
\description{
plot of normal
}
\examples{
##plot example5.6
x <- rnorm(9, sd = 2)
xx <- x- mean(x) + 20
exmp1 <- normnorm(xx, m = 25, s = sqrt(10), sigma = 2)
plot.normal(exmp1, leg_pos = "right", cex = 0.8)

## if mu and sigma  are unknown
y <- rnorm(13, mean = 2, sd = 0.5)
exmp2 <- normnorm(y, m = 0.5, s = 1, a = 2, b = 2)
## show the first plot : prior and posterior distribution of mu
plot.normal(exmp2, which = 1)
## show the second plot : prior and posterior distribution of tau
plot.normal(exmp2, which = 2, col = 1:2)
## show the third plot : Prior Contour
plot.normal(exmp2, which = 3, main = "Prior Contour",
           xlim = range(-3:3),ylim = range(0:3) )
## show the third plot : Posterior Contour
plot.normal(exmp2, which = 4, main = "Posterior Contour",
           xlim = range(-3:3), ylim = range(1:5))
}
