% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot.normal}
\alias{plot.normal}
\title{plot of normal}
\usage{
\method{plot}{normal}(x, y, leg_pos = c("topright", "bottomright", "bottom",
  "bottomleft", "left", "topleft", "top", "right", "center"), ...)
}
\description{
plot of normal
}
\examples{
##plot example5.6
x <- rnorm(9, sd = 2)
xx <- x- mean(x) + 20
exmp1 <- normal(xx, m = 25, s = sqrt(10), sigma = 2)
plot.normal(exmp1)

## if mu and sigma  are unknown
exmp2 <- normal(xx,m = 25, s = sqrt(10), a = 4, b = 1)
plot.normal(exmp2)
}
