% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poigamma.R
\name{poigamma}
\alias{poigamma}
\title{Poisson sampling distribution with a gamma distribution of prior}
\usage{
poigamma(x, shape, rate, scale = 1/rate)
}
\arguments{
\item{x}{the sample data from poisson distribution  ( \eqn{\g 0} ).}

\item{shape}{shape parameter for the gamma distribution ( \eqn{\g 0} ).}

\item{rate}{rate parameter for the gamma distribution ( \eqn{\g 0} ).}

\item{scale}{equals 1 / rate ( \eqn{\g 0} ).}
}
\value{
An object of class "\code{poigamma}" is returned.
}
\description{
Define the posterior distribution function for \eqn{\pi ( \mu | x )}, with
a gamma prior distribution \eqn{\pi ( \mu; \alpha, \lambda )} and a
poisson sampling distribution \eqn{L (\mu)}.
}
