% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_g12post.R
\name{summary.g12post}
\alias{summary.g12post}
\title{Summarizing binomial sampling distribution with a beta distribution of
prior}
\usage{
\method{summary}{g12post}(objects, ...)
}
\arguments{
\item{objects}{an object of class "\code{g12post}", usually, a result of a call
to \code{\link{binbeta}}.}

\item{...}{further arguments from other methods}
}
\value{
the function \code{summary.g12post} returns a list of summary
statistics of prior and posterior distribution, repectively.
}
\description{
A function used to produce summaries of prior and poesterior distribution,
based on binomial sampling distribution with a beta distribution of prior.
}
